file structure update
The git repo for notebooks is called notebooks-nextsim-workshop2025, so
just use this directly without another level of
nextsim-workshop2025/notebooks.

Check if this breaks anything in the notebooks? I think the paths
mostly are for data-nextsim-workshop2025 inside notebooks, so it should
be safe to change the location of the notebooks themselves.